+++ libcnb-data/src/lib.rs
@@ -1,10 +1,6 @@
-//! Low-level representations for Cloud Native Buildpack data types.
-
-// Enable rustc and Clippy lints that are disabled by default.
-// https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies
-#![warn(unused_crate_dependencies)]
-// https://rust-lang.github.io/rust-clippy/stable/index.html
+#![doc = include_str!("../README.md")]
 #![warn(clippy::pedantic)]
+#![warn(unused_crate_dependencies)]
 // This lint is too noisy and enforces a style that reduces readability in many cases.
 #![allow(clippy::module_name_repetitions)]