+++ libcnb-proc-macros/src/lib.rs
@@ -1,8 +1,6 @@
-// Enable rustc and Clippy lints that are disabled by default.
-// https://doc.rust-lang.org/rustc/lints/listing/allowed-by-default.html#unused-crate-dependencies
-#![warn(unused_crate_dependencies)]
-// https://rust-lang.github.io/rust-clippy/stable/index.html
+#![doc = include_str!("../README.md")]
 #![warn(clippy::pedantic)]
+#![warn(unused_crate_dependencies)]
 
 use proc_macro::TokenStream;
 use quote::quote;